airflow/requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-apache-spark
pyspark==3.5.1
pandas
psycopg2-binary
